requirements.txt
cffi==1.17.1
gitdb==4.0.12
GitPython==3.1.44
pycparser==2.22
smmap==5.0.2
